pre-hard code

+++ src/application/controller/controller_algorithm.h
@@ -11,23 +11,23 @@
 #define MIN_COOR_BOUND 0
 
 // output related const
-extern const float max_commanded_angle; // 10 degrees in radians
-extern float reference_signal; // no roll program for test flight
-extern const float commanded_angle_zero; // safe mode, init overwrite, p and c out of bound
+extern const double max_commanded_angle; // 10 degrees in radians
+extern double reference_signal; // no roll program for test flight
+extern const double commanded_angle_zero; // safe mode, init overwrite, p and c out of bound
 
 typedef union {
-    float gain_arr[NEW_GAIN_NUM];
+    double gain_arr[NEW_GAIN_NUM];
 
     struct {
-        float gain_k[NEW_ROLL_STATE_NUM];
-        float gain_k_pre;
+        double gain_k[NEW_ROLL_STATE_NUM];
+        double gain_k_pre;
     };
 
 } controller_gain_t;
 
-w_status_t interpolate_gain(float p_dyn, float coeff, controller_gain_t *gain_output);
+w_status_t interpolate_gain(double p_dyn, double coeff, controller_gain_t *gain_output);
 w_status_t get_commanded_angle(
-    controller_gain_t control_gain, float control_roll_state[NEW_ROLL_STATE_NUM], float *cmd_angle
+    controller_gain_t control_gain, double control_roll_state[NEW_ROLL_STATE_NUM], double *cmd_angle
 );
 
 #endif // CONTROLLER_ALGORITHM_H_
